+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The main API is stable, but expect some amount of breaking changes around major 
 - Add integrated self-hosted Swagger UI/ReDoc UI module
 
 ## 13.11.0 (uncut)
-- Upgrade to various dependency versions, including `Json4s` & `scala-xml`.
+- Upgrade to various dependency versions, including `circe`, `Json4s` & `scala-xml`.
 
 ## 13.10.1
 - Replaced `Circe.JsonFormat.patchBodySpec()` with `patchBody()` to ease use of Patch endpoint creations.